+++ AutumnEngine/Framework/Sprite.cpp
@@ -30,13 +30,13 @@ void AutumnEngine::Sprite::InitialiseQuad()
 	m_Quad[0].texCoords = sf::Vector2f(0, 0);
 
 	m_Quad[1].position = sf::Vector2f(GetPosition().x + GetSize().x, GetPosition().y);
-	m_Quad[1].texCoords = sf::Vector2f(GetTextureSize().x, 0);
+	m_Quad[1].texCoords = sf::Vector2f(static_cast<float>(GetTextureSize().x), 0);
 
 	m_Quad[2].position = sf::Vector2f(GetPosition().x + GetSize().x, GetPosition().y + GetSize().y);
-	m_Quad[2].texCoords = sf::Vector2f(GetTextureSize().x, GetTextureSize().y);
+	m_Quad[2].texCoords = sf::Vector2f(static_cast<float>(GetTextureSize().x), static_cast<float>(GetTextureSize().y));
 
 	m_Quad[3].position = sf::Vector2f(GetPosition().x, GetPosition().y + GetSize().y);
-	m_Quad[3].texCoords = sf::Vector2f(0, GetTextureSize().y);
+	m_Quad[3].texCoords = sf::Vector2f(0, static_cast<float>(GetTextureSize().y));
 }
 
 // DOESN'T NEED TO BE CALLED, SPRITE CAN BE RENDERED WITHOUT CALLING THIS DRAW METHOD
